requirements.txt
absl-py==2.1.0
accelerate==1.2.0
addict==2.4.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.7.0
astunparse==1.6.3
async-timeout==5.0.1
attrs==24.2.0
av==14.0.1
contourpy==1.3.1
crcmod==1.7
cryptography==44.0.0
cycler==0.12.1
dashscope==1.20.14
datasets==2.18.0
defusedxml==0.7.1
dill==0.3.8
distro==1.9.0
einops==0.8.0
exceptiongroup==1.2.2
flatbuffers==1.12
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.2.0
gast==0.4.0
google-auth==2.36.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.68.1
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.26.5
idna==3.10
jiter==0.8.0
jmespath==0.10.0
keras==2.9.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
libclang==18.1.1
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.9.3
mkl-service==2.4.0
modelscope==1.15.0
multidict==6.1.0
multiprocess==0.70.16
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.3.0
openai==1.57.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
oss2==2.19.1
packaging==24.2
pandas==2.2.3
pillow==11.0.0
platformdirs==4.3.6
propcache==0.2.1
protobuf==3.19.6
psutil==6.1.0
pyarrow==18.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyclipper==1.3.0.post6
pycocotools==2.0.8
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
PyJWT==2.8.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
qwen-vl-utils==0.0.8
regex==2024.11.6
requests-oauthlib==2.0.0
rsa==4.9
safetensors==0.4.5
scipy==1.14.1
sentencepiece==0.2.0
shapely==2.0.6
simplejson==3.19.3
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
supervision==0.21.0
sympy==1.13.1
tensorboard==2.9.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==2.5.0
tf-keras==2.15.0
tf-slim==1.1.0
timm==1.0.12
tokenizers==0.21.0
tomli==2.2.1
torch==2.5.1
torchvision==0.16.2
tqdm==4.67.1
transformers==4.47.0
triton==2.1.0
typing_extensions==4.12.2
tzdata==2024.2
websocket-client==1.8.0
Werkzeug==3.1.3
wrapt==1.17.0
xxhash==3.5.0
yapf==0.43.0
yarl==1.18.3
zhipuai==2.1.5.20241204
